+++ src/components/Messaging/MessageBody.js
@@ -22,7 +22,9 @@ import {
 import { TextInput } from './TextInput.js';
 import { MessagingContext } from 'context/MessagingContext.js';
 import SurveyCharts from './SurveyCharts.js';
-const log = require('loglevel');
+import * as loglevel from 'loglevel';
+
+const log = loglevel.getLogger('./MessageBody.js');
 
 const useStyles = makeStyles((theme) => ({
   messageRow: {
@@ -512,7 +514,7 @@ const MessageBody = ({ messages, messageRecipient, avatar }) => {
   const validateMessage = (payload) => {
     const errors = {};
 
-    console.log('body', /\w/g.test(payload.body.trim()));
+    log.debug('body', /\w/g.test(payload.body.trim()));
     if (payload.body.length === 0 || !/\w/g.test(payload.body.trim())) {
       errors.body = 'Please enter a message';
     }
@@ -539,7 +541,7 @@ const MessageBody = ({ messages, messageRecipient, avatar }) => {
     const errorsFound = Object.keys(errs).length > 0;
     if (errorsFound) {
       setErrors(true);
-      console.log('errors', errs);
+      log.debug('errors', errs);
     }
 
     if (messageContent !== '') {