+++ src/components/tests/organizations.test.js
@@ -44,11 +44,6 @@ describe('CaptureFilter organizations', () => {
         ReactDOM.unmountComponentAtNode(div);
       });
 
-      it('renders text "Verification Status" ', () => {
-        render(component);
-        expect(screen.getByText('Verification Status')).toBeInTheDocument();
-      });
-
       it('renders "Start Date" input ', () => {
         render(component);
         let input = screen.getByRole('textbox', { name: 'Start Date' });
@@ -95,7 +90,7 @@ describe('CaptureFilter organizations', () => {
         const orglist = screen.getByRole('listbox');
         const orgs = within(orglist).getAllByTestId('org-item');
         const listItems = orgs.map((org) => org.textContent);
-        console.log('default orgList', listItems);
+        log.debug('default orgList', listItems);
 
         expect(orgs).toHaveLength(2);
       });
@@ -143,7 +138,7 @@ describe('CaptureFilter organizations', () => {
         const orglist = screen.getByRole('listbox');
         const orgs = within(orglist).getAllByTestId('org-item');
         const listItems = orgs.map((org) => org.textContent);
-        console.log('default orgList', listItems);
+        log.debug('default orgList', listItems);
 
         // two default options + two orgs
         expect(orgs).toHaveLength(4);